Calibre: use system theme
Summary:
From the [Calibre manual](https://manual.calibre-ebook.com/customize.html):

> By default, on Linux, calibre uses its own builtin Qt style. This is to avoid crashes and hangs caused by incompatibilities between the version of Qt calibre is built against and the system Qt. The downside is that calibre may not follow the system look and feel. If you set this environment variable on Linux, it will cause calibre to use the system theme – beware of crashes and hangs.

As far as I understood, Solus' policy is to rebuild the children once a parent has been updated, so we'll never fall in the case specified in the manual.

Test Plan:
Installed the patched version. Breeze theme is correctly detected and used.
In Budgie, the theme is consistent as well.

Reviewers: #triage_team, DataDrake

Reviewed By: #triage_team, DataDrake

Subscribers: DataDrake, sunnyflunk

Differential Revision: https://dev.getsol.us/D4138